kernel: add async process loading trait
This isn't strictly needed at this moment, but one benefit of having an
async process loading framework is it would support loading processes
from other locations (e.g. an external flash chip). If someone were to
do that in the future, these traits would provide some structure for
doing so.