arch: add cortex-m4f crate
This adds a separate crate for ARM Cortex-M4F (aka Cortex-M4 with
FPU extensions). This is ARM lingo for hard float support. Currently,
Tock does not take advantage of hard float capability. This M4F crate
does not change that, it's basically a clone of the M4 crate, but it
allows us to differentiate which chips do have FPUs and is step zero
for adding hard float support.