boards/qemu_rv32_virt: unbreak QEMU >= 9.2.3, specify ELF file entry
Since somewhere around v9.2.3, QEMU has starting respecting the BIOS
ELF file header's entrypoint, which we're setting to 0x0 for all
boards. This caused QEMU to jump into the middle of invalid / unmapped
memory, thus busy-looping in trying to trap on the resulting illegal
instruction fault.

This adds the ELF file's entrypoint and sets it to the correct,
expected address, fixing this issue.